requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
config==0.5.0.post0
cycler==0.10.0
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
psycopg2-binary==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
rq==1.5.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
sqlalchemy==1.3.18
threadpoolctl==2.1.0
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
xlrd==1.2.0
